Refactor file structure and importing to absolute
