--- src/CreateBook.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import CreateBookFrom from './BookForm';
-
-const CreateBook = () => <CreateBookFrom />;
-
-export default CreateBook;
