+++ src/pages/CreateBook.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import CreateBookFrom from 'components/BookForm';
+
+const CreateBook = () => <CreateBookFrom />;
+
+export default CreateBook;
